Make Roda.rodauth a subclass of Auth instead of instance
Create an instance of Auth for each request that needs
authentication.  The instance of Auth has access to the
scope, which gives it access to the request and response,
so any method defined in Auth can use scope, request, and
response for specially handling per-request.

Make all configuration methods that accept values also accept
blocks, where the block defines the method directly, so any
configurable option can now return values that are based on
the request.

Add support for a logout feature, and make sure that login
clears the session for security.