Change Feature#require_login to #require_account
This ensures that there actual is a matching account and not a
bogus value in the session, and simplifies handling inside
the routes.

Change session_value to account_session_value, and add a
session_value that returns the current value of the session.

Add require_account and require_login auth methods, so they
are usable in routes.

Don't require logins to logout, since logout is really just
clearing the session.