Add ability to set behavior if visting page not requiring login when already logged in
In some cases, you don't want to allow this to happen.  By clearing
the session in update_session, it shouldn't be a security risk
to allow it, so by default it is allowed, but this way you can
disallow it if you want.

The logout path is special, since that shouldn't require a login,
but is generally used when logged in, so don't do any special
handling for it.