Remove overriding of route/get/post blocks
I'm not aware of this feature being used, and it added a lot of
complexity, mostly in terms of conceptual overhead. Referring to
these as auth block methods was confusing as both of the other
configuration methods also take blocks, and then operate in a
different context.

The feature isn't needed at all as you can just override the
routing blocks before calling r.rodauth.

Another change here is that Feature#route_block is being removed.
No internal features used it.  External features that can't use
get_block and post block should switch to setting the ROUTE_BLOCK
constant on the feature.