Switch unlock_account_autologin? to be true by default
If false by default, there is possible denial of service attack
preventing a user from accessing their account.  If the denial
of service attack could try an incorrect password more than 100
times between when the account was unlocked and when the user
could login, which seems probable in many cases.