Remove after_close_account from verify_account
The verify account key is cleared after the account is verified.
It shouldn't be possible to get to close account if the account
hasn't been verified.  So there's no reason to do this.  Worst
case scenario even if this did happen is that there would be
an extra row, which could probably be cleaned up after the
deadline passed.