Add correct security to the MySQL support
Previously, it was up to the user to set up security correctly on
MySQL, Rodauth only provided the functions.  This was because I
didn't know how to actually do it on MySQL.  However, I was able
to figure it out, so this has the migrations show users exactly
what rights need to be granted.

MySQL's grant/revoke support is fairly limited comapred to
PostgreSQL. It doesn't appear to allow you to grant access to
ALL tables by default, and then revoke access to a specific
table. On MySQL, it seems best to grant the password hash account
full access to the database, and give it the ability to grant
access to the application account. Then, each time the password
hash account creates a table that the application account needs
to use, it specifically grants access to that table to the the
application account.

This removes the permission change functions from rodauth/migrations.
It looks like the differences between PostgreSQL and MySQL make
this a poor abstraction.  Instead, just use a case statement in
the migrations.

This moves the MySQL setup/teardown SQL code to separate files
instead of inlining it into the Rakefile.  Also, it combines the
OTP migration with the default migration.