Add rodauth.param and _param
This abstracts the extracting of parameter values from the request.
Importantly, it handles the conversion to string, so that all
callers don't need to call to_s.  There are some places where
parameter values are used in if statements, so _param exists for
those, where it will return nil if the parameter doesn't exist,
but if the parameter does exist it will convert the value to a
string.