Check account status in rodauth.load_memory in remember plugin
Previously, if the account was closed, you could still login
with a remember token if there was a remember key.  This shouldn't
be possibly by default as closing the account removes the remember
key, but it's best to be sure.

In addition, this changes things to use update_session, so that
there is a central point for all places where the login id is
added to the session.