Add _before and _after internal methods, make ununderscored methods only for users
Previously if you overrode some before and after methods and didn't
call super, you could have had a bad time.  This makes it so you
never have to call super.

While here, handle cases where features work together without
strictly depending on each other.  Have both the _after and _before
hooks use "super if defined?(super)" so they will automatically
call a super method if one exists.  This is called before calling
the after/before public api methods.  If feature A wants to call
a before or after method defined by feature B, A just needs to
define a _before_B method and run it's own code, then use
"super if defined?(super)". This way, it doesn't matter in which
order the user loads the plugin.  Previously, if the user loaded
the plugin out of order, they would be missing some of the hooks.

To make it easier on users, document the before hooks in the
feature documentation, and remove the catch all language from
the README.

This also moves after_close_account from the base feature to the
close_account feature, as that is where it belongs.