Use timing safe comparison for comparing single session keys
A timing attack here should not be possible anyway, since an
attacker should not be able to inject values into the session.
However, since this is a comparison of keys, it's best to do
it for consistency with the other code.

Also, remove final else block, which was used if current_key =
nil.  If the single session key is set in the session, the
current key in the database should always be set, since it never
gets deleted.  However, on the off chance it is deleted, it's OK to
return nil instead of false.