Add convert_timestamp private method for converting timestamps
This DRYs up some code that manually called Time.parse.  The
only reason this is necessary is jdbc/sqlite returns the values
as strings, which is fair as SQLite doesn't actually support
timestamps.

Use DB.to_application_timestamp instead of Time.parse as the
conversion function.  In the specs, on jruby/sqlite set the
database timezone to :utc and the application timezone to
:local, so that the correct conversion will be done.