Attempt to handle unique constraint violations raised in race conditions where possible
Not all cases where a row is inserted can the unique constraint
violation be adequately handled, except by maybe an error page
(which the app should probably use).  However, for the cases where
such an error can be handled, attempt to handle it as best as
possible.

Currently, the unique violation handling code is untested, but it
is possible that it can be tested in the future.