Split otp_recovery_codes feature from otp feature
It's possible that some users will not want to allow recovery
code use, and it's easier to split the features that try to
add flags to disable parts of a feature.  Additionally, one
of Rodauth's principles is that you can only require what
you need.

Arguably, one could say that otp could be further split into
otp_auth, otp_setup, and otp_disable features.  I'm not sure
how possible that would be due to how intertwined the parts
are, or even if it would be useful. But it may be possible to
do that in the future, and just have the otp feature depend
on the other three features.