Make recovery codes and sms codes no longer depend on otp
This renames the otp_recovery_codes and otp_sms_codes to just
recovery_codes and sms_codes.  It makes otp, recovery_codes,
and sms_codes all depend on a two_factor_base feature, which
implements shared behavior.

The default table names for the recovery codes and sms codes
features have been changed, as well as all of the configuration
methods for the features.

While here, fix the down blocks in migrations to drop tables,
or in the travis case, just remove the down block.