Better timing safe comparison
Before we weren't making sure the provided string was the same
length as the actual string when doing the character by
character comparison.  This could lead to a timing issue in terms
of figuring out how long the actual string is.  As this method is
used with tokens, which are a fixed size by default, even that
minor bit of information is not likely to be helpful.