Add support for cached and private methods
This allows features to use auth_cached_method to create a
public instance method that caches the result from an underscored
private instance method, defining a configuration method with the
same name as the public method, which overrides the underscored
private method.

This also changes the hooks so that the public method is
not underscored, and the private method overridable by the
user is underscored, and the configuration method overrides the
private underscored method.