Execute get_block and post_block in the Rodauth::Auth instance scope
Virtually all of the lines in the routing blocks operated on the
auth variable, so it makes sense to just execute it in the context
of that object.  For the two factor authentication features that
use multiple routes, just assign request to r so the request
variable name remains the same.