Add verify_change_login feature for requiring account reverification on login changes
This moves the timestamp used for verify_account_grace_period to
the verify account table, as it wouldn't make sense to modify the
account creation timestamp if reverifying an account after a login
change.

This also does not allow changing the login for an unverified account.
This became a problem when verify_account_grace_period was added,
as the user could register an account with one email, get the
verification link, change the login, then verify the account using
the link, and they would have ended up with a verified account where
they didn't have access to the email.