Handle passwords that expire while the session is logged in
Previously, expiration was only checked and cached during login.
However, unless you were expiring sessions, the user could use
the current session indefinitely without changing their password.

This has the session keep a timestamp of when the password was
changed, so that if it expires during the session, the next request
will be redirected to the change password form.