Disallow possible reuse of TOTP authentication code
Previously, there was no check that the TOTP code was not already
submitted, which can lead to security issues in a passive MITM
attack (not much you can do in an active MITM attack).

We already had a last use column, but it was allowed to be NULL.
There wasn't a reason to allow NULL in this case, it was due to
an oversight, because the initial plan was to insert the secret
into the table before it was valid.  With the current design of
not inserting into the table until the authentication code has
been confirmed, it should never be NULL.

To avoid race conditions, due a proper check in the database, only
updating the last use column if it is more that the otp interval
from the current timestamp.  If it is within the otp interval,
there is a possibility the otp authentication code is reused, in
which case it is treated as an invalid code.