Fixes for some jwt feature issues
If :json=>:only plugin option is used, don't require that requests
have Content-Type application/json and raise an error if not, just
assume that Content-Type.

Make sure to set a response Content-Type of application/json for
json responses when using the jwt feature.

Don't assume the Authorization header contains the jwt token if
it starts with Basic or Digest, as those indicate HTTP Basic or
Digest authentication.