Add use_jwt? configuration method to jwt feature, for whether to use the JWT token or rack session for authentication information
This changes only_json? to only check request content types for
rodauth endpoints, not for all endpoints.  For non-rodauth endpoints,
instead of checking the request's content type to determine where to
get the session information, default to using the Authorization
header if present.  If the Authorization header is not present,
then assume an empty session if in only_json? mode or when the
request uses a json content type. The user can use the use_jwt?
configuration method to change the criteria, which may be necessary
in certain cases.