Use 4xx error statuses for errors, instead of using a 200 success status
Using 200 as a status for errors was not intentional, that was
just the default because Roda defaults to that for responses with
bodies.  Browsers don't treat 200 different from 4xx, so this
shouldn't affect browser users.

JSON API (jwt feature) users this would affect, as previously
400 was used for all responses. For backwards compatibility, the
400 behavior for JSON API responses is kept, but it can switch
to the newer 4xx statuses via the json_response_custom_error_status?
configuration method.  Rodauth 2 will default to using 4xx statuses
for JSON API responses.

The specs set json_response_custom_error_status? to true to
start checking for the expected 4xx status.