Fix resending verify account email when attempting to create a new account with same login as unverified account when using verify_account_grace_period feature
This adds an allow_resending_verify_account_email? method that
controls whether resending the verify account email should be
allowed.  By default, this is only if the account is currently
in unverified state.

Previously, if the account was in the unverified grace period,
open_account? would be true, so you couldn't resend the email.

Additionally, in new_account, don't show them the verify account
resend view if they wouldn't be able to resend if clicking the
button.