Add verify_login_change feature as an alternative to verify_change_login, where the change doesn't take affect until after verification (Fixes #31)
The verify_change_login feature changed the login and required
verification, which can result in the user being locked out of
their account if they used the wrong email.  This doesn't change
the login on the account until after verification.

Unfortunately, verify_login_change is a much more complex feature
than verify_change_login, because it can't just piggyback on the
verify_account support.

This switches the demo-site to use verify_login_change instead
of verify_change_login, and updates the documentation for
verify_change_login to note that it is deprecated.