Deal better with time differences between the database and application servers in the password_expiration plugin
At least when running specs, if there are time differences between
the database and application servers, the default of 0 for
allow_password_change_after may not always allow password changes.
Set it to -86400 so the database server can be at least a day
behind.

Minor spec changes to give greater tolerance when the database server
is significantly ahead.

This could be made a little more robust by always using the
database's current timestamp, but that would probably hurt performance,
and in practice the database and application servers time is not
going to differ so much as to matter.  This is mostly just an
issue when testing.