Add require_http_basic_auth configuration method to http_basic_auth feature (Fixes #41)
On sites that only use rodauth for the basic auth feature,
redirecting to a login page may not be helpful.  Try to support
that behavior using the require_http_basic_auth method, which
will make it so require_authentication will return a 401 response
instead of a redirect to the login page.