Add create_account_set_password? and verify_account_set_password? methods to delay setting password until account verification
This can fix a possible issue where an attacker creates an account
for a user with a password the attacker knows.  If the user clicks
on the link in the verify account email and clicks on the button on
the verify account page, the attacker now has a verified account
that they know the password to.

With this new approach, only the user who has access to the email
can enter the password for the account, which should fix that
issue.  This new approach may become the default in Rodauth 2.