Add disallow_common_passwords feature, for disallowing the usage of the most common passwords
This uses a list of 10,000 most common passwords taken from the OWASP
SecLists project.  Users can supply their own file, but I wanted to
include a least a basic file that is usable.  I originally wanted to
include the 100,000 most common passwords, but that file was much
larger than the entire gem, so I settled on the 10,000 most common
passwords.