Support Roda route_csrf plugin for request-specific CSRF tokens
Roda's new route_csrf plugin allows for request-specific CSRF
tokens, which are much more secure than the tokens used by
the csrf plugin (which uses rack_csrf).

The route_csrf plugin is not a drop-in replacement, however,
so it is not being turned on by default.  It will probably
become the default in Rodauth 2.

Unless this support is enabled via the :csrf=>:route_csrf
option, this should be backwards compatible, except that
the csrf_tag method now takes an optional path argument.

You can run specs with the RODA_ROUTE_CSRF=1 environment
variable set to test this new support. You can run specs
with the RODA_ROUTE_CSRF=2 environment variable to test
this support without turning on support for request
specific tokens.