Work with Roda sessions plugin, using string keys for session information if that is used
This includes all necessary changes to work with the sessions
plugin.  There is a new session_keys features method, similar
to auth_value_method, but where the default implementation is
to convert the value to a string if the Roda
:sessions_convert_symbols option is set.

There was one place where a symbol key was hardcoded, in the
confirm_password feature for the key storing the redirect.
Add a configuration method for that.

If the scope has a clear_session method, then that is used
to clear the session, as opposed to a simple session.clear,
for better integration with Roda's sessions plugin.