Add email_auth feature, for authentication using links sent via email
This allows for passwordless authentication.  For sites where
usage is infrequent it can make more sense to just email a login
link instead of requiring a password.

While this feature supports accounts without passwords, it still
works with accounts that have passwords, allowing the user to
login via password or via a login link in that case.