Gracefully handle failure when new login is already taken in the verify_login_change feature
Previously, verify_login_change would accept a login for an
existing account.  When you tried to verify the account, the
database uniqueness constraint would raise an error, which Rodauth
would not handle.

This adds checking in two places.  When you use the change login
form, it will check for an existing account with the new login,
and fail if the account already exists, similar to the use of
the change_login feature without verify_login_change.  When
verifying accounts and actually changing the login, uniqueness
errors are now caught and reported similar to how it is handled
in other features.