Add default_post_email_redirect configuration method, setting default for all redirects after emailing when not logged in
Before, the redirect behavior after emailing if not logged in
was a little inconsistent.  Most features used the default_redirect,
but verify_account used the require_login_redirect. This was
compounded by the recent addition of rate limiting, which used
a different redirect if the email was not sent due to rate
limiting.

This allows you to set the behavior in one place.

The only backwards incompatibility change from the last release
is that verify_account_email_send_redirect now goes to /
instead of the /login by default.