Add otp_setup_hmac_secret configuration method to otp feature, used to ensure TOTP secrets were generated by Rodauth
This allows Rodauth to enforce that the secret submitted is one
generated by Rodauth and not an attacker.  The previous behavior
allowed the user to submit any valid OTP secret, and in the case
of the JSON support, that was actually required.

JSON support for this is kind of rough.  You first post to the
otp-setup route, and that will fail but provide the secret and
hmac to use, and then you can post again with the secret and
hmac and it will succeed.