Add support for accessible error handling
This makes error handling accessible by default, using aria-invalid
for the fields with errors, and aria-describedby to tie the input
field to the error message.

This also adds support for making the input fields as required, so
that browsers will not allow submission of the form without the
fields.  That change is not on by default as it modifies behavior,
but you can turn it on using mark_input_fields_as_required?.

To add more support for customization, this adds the following
configuration methods:

* formatted_field_error
* field_error_attributes
* field_attributes
* default_field_attributes
* input_field_error_message_class

To DRY up template code, this also adds input_field_string, which will
return an HTML fragment for the input, including an error message
associated with it if necessary.