Use urlsafe base64 for new token keys on Ruby 1.8
Ruby 1.9+ has always used urlsafe base64, so this makes the
algorithm the same across Ruby versions.

Ruby 1.8 SecureRandom didn't have a method for this, but it is
possible to emulate using pack/chomp/tr.

This will not affect existing tokens in the system, but it makes
new token keys shorter on Ruby 1.8 (43 bytes instead of 64).