Add email_token_hmac_secret configuration method to HMAC tokens sent to users
This method turns on the HMACing of tokens.  By sending users an
HMACed token and not a raw token, and only allowing login using
an HMACed token, this makes it so even if the raw tokens are leaked
(e.g. via an SQL injection vulnerability), the raw tokens cannot
be used to gain access, unless the attacker also has knowledge of
the HMAC secret.

To allow for a graceful transition to HMACed tokens, an
allow_raw_email_token method is added.  If set to true, then
raw email tokens are allowed.

This is added to the email_base feature, so it affects all features
that use email_base to send emails:

* email_auth
* lockout
* reset_password
* verify_account
* verify_login_change

All of these except verify_account have a 1 day expiration in the
recommended migration and default interval, so the transition
period should be short.

verify_account does not have an expiration, so after the
allow_raw_email_token setting is set back to false, users will need
to rerequest a verify account email to get an HMACed token in order
to verify their account.