Combine otp_setup_hmac_secret and email_token_hmac_secret into hmac_secret
This adds an hmac_secret configuration option to the base, and has
the email_base and otp features that use HMACs switch to using
it.  We haven't shipped a version with the otp/email_base HMACs
enabled, and it is better to make this change before doing so.

While here, add documentation for HMACs to the README.