Use HMACed OTP keys if hmac_secret is set
If hmac_secret is set, this changes the otp to make OTP keys provided to
the user to use an HMAC, and keeps the raw OTP key in the database. This
way, if the OTP keys from the database leak, they are not usable for
two factor authentication unless the hmac_secret is also compromised.

This changes how hmacs were used during otp setup.  Now, both the
raw and HMACed OTP key are provided during setup, and the system
checks that HMACing the raw key results in the submitted HMACed key,
still ensuring that the key provided was randomly generated by the
server.