Log user out when locking out OTP authentication if no fallback options available
If no fallback options are available, there isn't much the user
can do.  It would be possible in some configurations to continue
in logged in but not fully authenticated mode, but in common
configurations it would not be possible.

The previous behavior in the committed spec is a redirect loop,
which is definitely not desireable.