Do not add auth methods for *_path and *_url
These should not be overridden separately from the *_route method.
This reverts the changes to the docs to add the methods.  The docs
document the configuration methods, not all methods available on
the rodauth instance.  It might be useful to mention the *_path
and *_url methods somewhere in the documentation, maybe in the
README.

Remove use of login_path by base feature, since login feature may
not be used, fixing tests.