Add #valid_login_entered? method for multi phase login
When using the email_auth feature, we want to display the option to
authentication via email when the user has entered a valid login during
multi phase login.

By default, the email_auth feature overrides #login_form_footer to add
the form for requesting email authentication. However, in my case I'm
rewriting Rodauth views in Rails, and don't want to use Rodauth's view
helpers such as #login_form_footer.

So, we add the #valid_login_entered? method that can be called in the
views, like this:

  # ... login form ...

  if rodauth.valid_login_entered?
    rodauth.email_auth_request_form
  end

We also get rid of the @email_auth_request_form variable which doesn't
seem to be necessary. This changes the behaviour to also display the
email auth request form when the user has entered both login and
password, but the password was invalid, which I think is desirable.