Make rodauth.authenticated? return true on OTP setup page (Fixes #68)
On the OTP setup page, OTP has not been setup yet, but the user must
have logged in, so authenticated? should be true.  It was only false
because Rodauth was treating the temporary otp key created during
setup as a permanent otp key for the purposes of determining whether
the user had an existing OTP key.

Note that this only affects the OTP setup page, as that is the only
page generating temporary OTP keys.  Also, authenticated? is not
called in the Rodauth's code on this page, so this can only affect
people calling authenticated as part of the layout or view or when
overriding methods that are called.

In cases where all users are required to have an OTP key, this
change may not be desired.  However, such cases should probably
check both authenticated? and two_factor_authentication_setup?
instead of assuming the previous behavior.