Drop Ruby 1.8 support
While there isn't much Ruby 1.8 specific support in Rodauth,
there is some, and it's unlikely anyone is using the Ruby
1.8 support.

Make the README reflect the new Ruby support policy, which
is that supported Ruby and JRuby versions are supported.
Unsupported Ruby and JRuby versions may be supported, but
it's possible support for them could be dropped in any
minor version. Currently, Ruby 1.9.2 is the minimum supported
version.