Support WebAuthn on demo site if RODAUTH_WEBAUTHN environment variable set
WebAuthn only works in the browser in https mode, and the public
demo on Heroku doesn't use https, and most people testing the demo site
probably won't have https enabled, so it doesn't make sense to enable
by default.

Also, remove the password/account/session expiration features
on the demo site, as they make the demo site annoying to use.

Add a not found and error handler for a nicer experience.

Use https for the bootstrap CSS, needed when testing the demo
site with https enabled.