Change default setting of verify_account_set_password? in verify_account feature to true
This is a more secure setting, since the person with actual access
to the email sets the password.  With a setting of false, the
password could be set by an attacker, if they can convince the
email account owner to perform the verification.

However, when the verify_account_grace_period feature is used,
the password should be asked for up front because the account is
allowed to login before verification.