Do not require login confirmation by default if verifying accounts or login changes
I was originally thinking of adding separate settings for the
verify account and verify login change cases, but doing so doesn't
make much sense.  If you aren't using both, you don't need two
settings, and if you are using both, it seems very strange to want
login confirmation in one case and not the other.  In any case,
if you do want that, you can pass a block to the method and look
at the request path to get different behavior.