Store authentication methods used in the session, available via rodauth.authenticated_by
Previously, Rodauth stored the authentication method used for two
factor authentication, but not for single factor authentication.
Rodauth now stores an array of all authentication types used in
the session.  Currently, this should be limited to 1 or 2 elements.
Mostly this is done for simplification, but potentially it opens
the door to 3-or-more factor authentication support in the future.

Since there are many features that support autologin, also support
a way of determining which type of autologin was used.

For the confirm_password feature, when the password is confirmed,
if autologin, remember, or email auth was used to authenticate,
switch the authenticated by to password instead.

Previously, Rodauth used symbols for the two factor auth type, but
this standarizes on using strings for the authentication type, to
make it simpler to support sessions that do not handle symbols.

Remove the clear_remembered_session_key and two_factor_session_key
configuration methods, as those session keys are no longer needed
after this change.