Add webauthn_login feature for passwordless login via WebAuthn
This builds on top of the webauthn feature to support webauthn
authentication as the primary or sole authentication support.

Currently, in order to setup WebAuthn authentication, you need
to already be authenticated, so truly passwordless authentication
would require either email auth or create/verify account autologin.

This makes some minor modifications to the webauthn feature:

* Allow setting up and removing webauthn authenticators if logged in
  via webauthn but not two factor authenticated.

* Do not show webauthn as a two factor authentication link if
  logged in via webauthn but not two factor authenticated.

* Add webauthn_account_id, which defaults to session_value, but can
  be set to account_id during login (since the session_value isn't
  set at that point).

* Add some minor changes to the webauthn-auth form to get it to
  work with webauthn-login.