Allow confirm_password feature to operate as second factor authentication if using webauthn login
This refactors the confirm password handling to add password if it
was not already present as an authentication type.

As confirm_password does not depend on two_factor_base, this
requires the (super if defined?(super)) pattern.

This changes the confirm_password form to only require single
factor authentication.  The form doesn't make any database
changes, only confirms the existing password, and I don't think
it makes sense to require two factor authentication for it.