Remove session keys instead of setting them to nil
Introduce remove_session_value and use this both for cases where
session.delete is used and where session[key] = nil is used.

Similar to set_session_value, override in the jwt feature to set
the jwt token.  None of the specs appear to require this, but
for consistency it makes sense to set the new JWT whenever the
session changes.