Allow redirecting to originally requested URL after requiring authentication
When the user visits a certain page, but this page requires
authentication, it's nice if the user is redirected back to the
originally requested page after successful authentication.

We implement this by storing the current URL into a session on
`login_required` and `require_two_factor_authentication`, and then
redirecting back to that URL on successful authentication, removing
the session value in the process.

This works with email authentication and webauthn login as well. We use
separate settings for login and two factor authentication, as two factor
authentication doesn't require the login feature to be loaded, and the
user might want to configure these separately anyway (e.g. redirect to
two factor auth directly after login, but have two factor authentication
remember requested URL).

We don't override `login_redirect` and `two_factor_auth_redirect` so
that they can still be used for configuring default redirects for when
the user manually visited the authentication page, while still keeping
the new functionality.

This feature is disabled by default because it could lead to unexpected
redirects in a scenario like the following:

1. user browses a website
2. they encounter a page that requires authentication
3. they go back an continue browsing the website some more
4. at some point they decide to manually login in / create account
5. they're redirected to the previously stored URL that was never cleared

This decision also allows to avoid breaking existing tests that aren't
counting with this functionality.

I considered solving this functionality using flash messages that would
be discarded when user continues browsing the app, and thus would reduce
the chance for an unexpected redirect. However, it would introduce
backwards compatibility for users iterating over all flash keys assuming
they're all either message that need displaying. It also introduces
complications with the JWT feature which doesn't load flash by default.