Don't display reset password request form in login header
We're already displaying a "Forgot Password?" link in the login footer,
so if the user didn't successfully enter their email address, they will
probably see that link and follow it to the reset password request form.

This removes the need for a hidden login field, which allows us to add
validations to the form, allowing the user to re-enter the login if it
was mistyped.