Remove confirm_password dependency for remember feature
The confirm_password feature was originally part of the remember
feature, but now that it's extracted, it appears these two features are
actually decoupled from one another.

While it provides extra security to require password confirmation for
sensitive parts of the app when logged in via remember token, for those
that aren't doing this it's beneficial if they don't need to load the
confirm_password feature code.