Rename require_password_confirmation to require_password_authentication (Fixes #75)
The problem with require_password_confirmation as a name is we
already have require_password_confirmation? which means something
different (whether to ask for passwords twice when setting them).
Also, if you haven't entered your password yet, it's more
authentication than confirmation.

Additionally, change the method so that it doesn't always redirect.
By default, it redirects if the session was not authenticated via
password.  The password_grace_period feature overrides it so that
it will also redirect if the password was not entered recently.

Rename some related methods to use a
password_authentication_required_* naming scheme.

Make require_password_authentication also require a login if the
user has not yet logged in.