Add translate(key, default_value) configuration method and have it affect all translatable content
This should take care of most translation needs.  Email bodies
will still require their own translation approach, but as those
are not static, they cannot use the simple approach used here.

Implement this by adding a translatable_method to features. This is
an equivalent to auth_value_method, except the method it defines
calls translate with the two arguments.  Default translate to
return the second argument, so other than the indirection, this
results in the same behavior.

In password_complexity, to make the too many characters error
message translatable, use a generic static message instead of
a more specific message.