Automatically check CSRF tokens before calling any Rodauth route by default, allow disabling using check_csrf? false
The Rodauth 1 default of using the Roda csrf plugin handled CSRF
checking automatically, since it was handled by the middleware.
With the switch to using the route_csrf plugin by default,
Rodauth routes were no longer checking CSRF tokens.

This detects when the route_csrf plugin is being used, and
automatically checks the CSRF token before dispatching to the
Rodauth route.

This makes the application :rodauth_csrf option to be the same
as the :csrf option given to the plugin, and makes the
application :rodauth_route_csrf option be true or false depending
on whether the route_csrf plugin is used.
Remove the adding of check_csrf! in the specs, since Rodauth
itself now handles this.  This shouldn't effect non-Rodauth
routes in the specs, since I believe the only POST requests
in the specs are to Rodauth routes.