Specify webauthn auth form action directly
This makes it easier for others to reimplement the webauthn-auth
template in other templating language and other web frameworks.

For example, in Rails, forms are created with a `form_*` helper method,
which automatically adds things like the CSRF token. There it's not
really possible to reuse a raw list of HTML attributes (unless there is
parsing involved).

It seems that the `webauthn_auth_form_attr` setting is only used by
webauthn_login feature to set the form action, and since webauthn_login
already overrides `webauthn_auth_form_path` with the same value, we can
use that directly instead.