Enable email rate limiting by default in lockout, reset_password, and verify_account features
This wasn't enabled by default when support for it was introduced
in 1.19.0, because it requires you add a column to the database
table in order to handle the last sent time.  However, it is a
better default.  The settings can be switched back to nil if the
database column can't be added.