Add audit_logging feature, logging changes to a database table
This is implemented with a new hook_action configuration method,
called for all hook calls.

One intresting aspect of the audit_logging feature is it is the
first feature to use configuration_module_eval for custom method
definitions.  I thought this approach would be easier and more
flexibility that an approach that used a hash mapping action
symbols to log messages, for example, plus it is similar to how
auth value methods work in regards to taking either a value or
a block.